Add linter for json, javascript and solidity

+++ contracts/levels/BankFactory.sol
@@ -1,12 +1,11 @@
 pragma solidity ^0.4.18;
 
-import './base/Level.sol';
-import './Bank.sol';
+import "./base/Level.sol";
+import "./Bank.sol";
 
 contract BankFactory is Level {
-
-  uint supply = 14657921;
-  uint playerSupply = 29;
+  uint256 supply = 14657921;
+  uint256 playerSupply = 29;
 
   function createInstance(address _player) public payable returns (address) {
     Bank bank = new Bank(supply);
@@ -14,7 +13,10 @@ contract BankFactory is Level {
     return bank;
   }
 
-  function validateInstance(address _instance, address _player) public returns (bool) {
+  function validateInstance(address _instance, address _player)
+    public
+    returns (bool)
+  {
     Bank bank = Bank(_instance);
     return bank.balanceOf(_player) > playerSupply;
   }
